n-dimensional (sub)hyper-cubes, n=0..6
The 4x4x4 6-D Hypercube is optionally (by selecting A12) divided up
into a 2x4x4, a 2x2x4, a 2x2x2, a 1x2x2, a 1x1x2, and 2 1x1x1 'cubes'.
The 'blinking' dimension is 6-D, and can be used to move the state to
the smaller cubes within.